feat: editor분리 및 컴포넌트화
